[bug 6063036] - Updated hadooptest.TestSession to allow the user to specify an alternate location and name for the framework configuration file, from the default of userhome/hadooptest.conf.
If the hadooptest.config system property is set, then the framework will
attempt to use that specified location for the framework configuration
file.

If the hadooptest.config system property is not set, then the framework
will use a default of userhome/hadooptest.conf, where "userhome" is the
path to the user's home directory (like /homes/rbernota/) and
"hadooptest.conf" is the default name of the configuration file.

To specify the alternate location from the command line, you can specify
the hadooptest.config to the JVM as follows, with the -D option:

-Dhadooptest.config=/alternatepath/alternatefilename

As an example when running a JUnit test, it would be like:

/home/y/bin/java -cp $CLASSPATH
-Dhadooptest.config=/homes/rbernota/framework_config/framework.conf
org.junit.runner.JUnitCore MapredKillTask

...hadooptest.config value will be picked up by the JVM and is read by
hadooptest.TestSession to specify the alternate location of the
framework configuration file.

If the -Dhadooptest.config= was not specified, then the framework would
assume the default location (userhome/hadooptest.conf).

As part of this changelist, I also made a quick fix to errant file
references in the .project file for hadooptest.