[bug 6042798] - Modified killTaskAttempt in hadooptest.job.Job to use the Hadoop API.  Added kerberos security support for all Hadoop API in Job that uses the JobClient.  Added more default configs to the default TestConfiguration.
Started adding an API for launching sleep jobs though the Hadoop API.

Stared correcting a problem with task attempt killing where I was trying
to get the task attempt status before the job was running.  I now wait
for the job to get out of prep state before trying to kill the task
attempt.

Fixed a problem in Cluster where the cluster was successfully in safe
mode, but the safe mode check was checking for a string to equal the
safe mode on string, when it should check if it is simply contained.