[bug 6042798] - Modified killTaskAttempt in hadooptest.job.Job to use the Hadoop API.  Added kerberos security support for all Hadoop API in Job that uses the JobClient.  Added more default configs to the default TestConfiguration.
Fixed a problem in Cluster where the cluster was successfully in safe
mode, but the safe mode check was checking for a string to equal the
safe mode on string, when it should check if it is simply contained.

Added a number of methods to Job to get status of tasks, to get the
JobClient for the job, to get the RunningJob, and JobID (Hadoop API).