[bug 6733858] - Ported GDM secure remote DFS cluster action code from branch-0.23.  Notes follow below.
-----------------------

Working final code for copying a file to, deleting, and
checking file status for a secure, remote Hadoop DFS from a GDM node,
using either hadoopqa or dfsload users, and using either HDFS or WEBDFS.

See the test hadooptest.gdm.regression.TestGDMRemoteDFSAccess for usage
of new APIs added.

Created a new dfs package hadooptest.cluster.hadoop.dfs.

Added DFS classes DoAs and PrivilegedExceptionActionImpl to enable
connecting to secure remote DFS's.

Added a number of public API's to the DFS class, to enable copying a
local file to a secure remote DFS, deleting that file, checking if the
file exists on that DFS, creating a dummy local system file for testing,
getting the UGI for a user connecting to a secure remote DFS, getting a
basic configuration for connecting to a secure remote DFS, and other
supporting methods.