[bug 6816290] - Added functionality to allow Storm tests to get Igor role info.
Added following APIs to StormCluster:

- lookupIgorRole(roleName)
- lookupIgorRoleClusterAllNodes(clusterName)
- lookupIgorRoleClusterContrib(clusterName)
- lookupIgorRoleClusterDrpc(clusterName)
- lookupIgorRoleClusterNimbus(clusterName)
- lookupIgorRoleClusterRegistry(clusterName)
- lookupIgorRoleClusterSupervisor(clusterName)
- lookupIgorRoleClusterUI(clusterName)

Converted StormCluster from interface to abstract class, so that
reusable code can be put there for both LocalModeStormCluster and
YahooStormCluster (and any other cluster type implemented for Storm in
the future).  Changed all StormCluster interface methods to abstract.

Mode LocalModeStormCluster and YahooStormCluster extend StormCluster,
instead of implement the old interface.

Changed ModifiableStormCluster from an interface to an abstract class,
since it now has to extend StormCluster, which is now an abstract class
(instead of an interface).