[bug 6857051] - Ported Spark YARN-client support to master branch for Hadoop 2.
-----

Updated the SparkHdfsLR App type to be able to support
submitting jobs in YARN-client mode.

Added tests to TestSparkHdfsLrCli to test submission and completion of
Spark Apps submitted with both YARN-client mode and standalone mode.

Additionally fixed a long-standing bug in the various job submit()'s
where the BufferedReader was never being closed on the process when the
job ID was found.  This was causing a problem with the YARN-client mode
app submission, because it was halting the YARN-client mode process when
we found what we wanted from the BufferedReader, but left it attached
when walking away from the process.

-----

Added an enumerated type to handle the master type
(yarn-client or yarn-standalone).

-----

Added YARN-client support for Spark App type SparkPi.

Also added another test in TestSparkPiCli that runs the same test,
additionally in YARN-client mode.