[bug 6335755] - Completely reworked dependencies for HTF.
- Eliminated maven offline mode.
- Eliminated dependencies directory.
- Dependencies on other products are now "yinst fetch"'ed and unpacked
to set up system dependencies to compile against on the grid.
- Set up a series of new maven profiles to support running and compiling
against all of the different products.
- Heavily modified parent-pom.xml and pom.xml.
- Added a number of explicit dependencies in the maven poms.
- Modified some framework classes and tests to support Hadoop 2.4 API.

As of this commit, Hadoop 2 and Spark clusters are verified compiling
and running tests successfully with HTF.  

Also as of this commit, local compile on a corp box is verified.

Storm and GDM are still
unverified.