[bug 6335755] - Eliminated parent-pom and pom-ci and merged config into a single master pom.xml.
Added profile profile-all to be able to maintain classpath dependency
order with the basic dependencies.

Added Maven module "topologies" for Storm topologies.

Added Maven module "htf-common" for the core HTF.

Various fixes and adjustments for compile and runtime with the new Maven
module reactor architecture.

As of this commit, it represents a working HTF with maven modules that
has been runtime tested against hadoop 2, spark, and storm.