Added explicit dependencies for log4j in maven profile for profile-storm and profile-ci.
There was an update to the ystorm_contrib package that probably included
an incompatible version of log4j.  This would have been showing up in
our classpath prior to our existing log4j dependency.  Adding explicit
log4j dependencies at the top of the profiles ensures that we'll always
be getting the log4j in HTF that we want.